Don't hide token code input, there's no need
Tokens are one time passwords valid for a very short time window and it
seems standard practice to echo them as entered (e.g. AWS console), as
the risk of exploiting them is low, yet hiding them negatively impacts
the user interface and can complicate code that validates user input.

So, let's make things simple and echo the token as entered, and have
awsmfa behave like the AWS console, not like a unix password prompt.

Note: code should work with python 2 and 3, but only tested on 3.7.5